backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
anthropic>=0.39.0
